[#936] Extract synthetic and tonemap modules (item 9)
Split the synthetic cLUT profile builders (create_RGB_A2B_XYZ,
create_synthetic_clut_profile, create_synthetic_smpte2084_clut_profile,
create_synthetic_hdr_clut_profile, create_synthetic_hlg_clut_profile) out of
icc_profile/__init__.py into icc_profile/synthetic.py, and the multiprocessing
tonemap/black-point-blend workers (_blend_blackpoint, _mp_apply,
_mp_apply_black, _mp_hdr_tonemap) into icc_profile/tonemap.py.
icc_profile/__init__.py re-exports the public builder functions plus
_mp_apply (both have confirmed external consumers in worker.py); the other
underscore-prefixed tonemap helpers have none and stay internal to
tonemap.py.

Resolves item 3's deferred import: tags/lut.py's LUT16Type._apply_black()
now imports _mp_apply_black directly from DisplayCAL.icc_profile.tonemap
instead of from DisplayCAL.icc_profile, since tonemap.py has no dependency
back on the tags package.
